fix: pass drift conditions through TopologyDriftUpdated event
The on_topology_drift hook was firing with a hardcoded generic message
("drift_detected") instead of the specific drift type (e.g. "missing_node").
This happened because TopologyDriftUpdated only carried a Bool flag,
discarding the drift conditions computed in Worker.hs.

Now the event carries [DriftCondition], and applyEvent fires per-condition
hooks with the correct drift type and details. DriftCondition and
renderDriftCondition are moved to Types.hs as shared definitions, and the
now-redundant decideDriftActions is removed.
